스케줄 관리자 기능 구현

+++ src/main/java/com/apartment/apart/domain/report/ReportController.java
@@ -32,7 +32,7 @@ public class ReportController {
     }
     @PreAuthorize("isAuthenticated()")
     @GetMapping("/detail/{id}")
-    public String detail(Model model, @PathVariable("id") Integer id, ReportReplyForm reportReplyForm) {
+    public String detail(Model model, @PathVariable("id") Long id, ReportReplyForm reportReplyForm) {
         Report report = this.reportService.getReport(id);
         model.addAttribute("report", report);
         return "report_detail";
@@ -55,7 +55,7 @@ public class ReportController {
 
     @PreAuthorize("isAuthenticated()")
     @GetMapping("/modify/{id}")
-    public String reportModify(ReportForm reportForm, @PathVariable("id") Integer id, Principal principal) {
+    public String reportModify(ReportForm reportForm, @PathVariable("id") Long id, Principal principal) {
         Report report = this.reportService.getReport(id);
         if(!report.getUser().getUserId().equals(principal.getName())) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
@@ -68,7 +68,7 @@ public class ReportController {
     @PreAuthorize("isAuthenticated()")
     @PostMapping("/modify/{id}")
     public String reportModify(@Valid ReportForm reportForm, BindingResult bindingResult,
-                                  Principal principal, @PathVariable("id") Integer id) {
+                                  Principal principal, @PathVariable("id") Long id) {
         if (bindingResult.hasErrors()) {
             return "report_form";
         }
@@ -82,7 +82,7 @@ public class ReportController {
 
     @PreAuthorize("isAuthenticated()")
     @GetMapping("/delete/{id}")
-    public String reportDelete(Principal principal, @PathVariable("id") Integer id) {
+    public String reportDelete(Principal principal, @PathVariable("id") Long id) {
         Report report = this.reportService.getReport(id);
         if (!report.getUser().getUserId().equals(principal.getName())) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이 없습니다.");
@@ -93,7 +93,7 @@ public class ReportController {
     @PreAuthorize("isAuthenticated()")
     @GetMapping("/like/{id}")
     @ResponseBody
-    public String reportLike(Principal principal, @PathVariable("id") Integer id) {
+    public String reportLike(Principal principal, @PathVariable("id") Long id) {
         Report report = this.reportService.getReport(id);
         SiteUser siteUser = this.userService.getUser(principal.getName());
         this.reportService.like(report, siteUser);
