+++ src/main/java/com/apartment/apart/domain/reportReply/ReportReplyController.java
@@ -26,7 +26,7 @@ public class ReportReplyController {
     private final UserService userService;
     @PreAuthorize("isAuthenticated()")
     @PostMapping("/create/{id}")
-    public String create(Model model, @PathVariable("id") Integer id,
+    public String create(Model model, @PathVariable("id") Long id,
                          @Valid ReportReplyForm reportReplyForm, BindingResult bindingResult, Principal principal) {
 
         //답변 부모 질문객체를 받아온다.
